Add files via upload
Restored "blob filter" to TPU AI, not sure why I'd removed it, with spring weather spiders and bugs moving in front of the cameras proved its needed.  Default theshold of 0.20, still seems fine for my opservations even with 4K cameras.  These oub of focus "blobs" detecting as a person with high probability is perhaps the single largest problem with the Mobilnet-SSD v1 & v2.