Add files via upload
When I removed fisheyeTPU.py and  renamed fisheyeTPU.py as TPU.py I forgot that the "blob filter" was not installed in fisheyeTPY.py  Added the blob  filter back into TPU.py.  The blob filter rejects a detection that covers some % of the entire frame (default 20%).  This is a heuristic developed to reject false detection caused insects attracted to the camera IR lights flying close in front of the cameras.